Move backend/locations -> core/locations
